Fix fed1 schema upgraded when `@external` is on a type. (#2343)
For historical reasons, `@external` was legal to put on an object type
since about forever (even in federation 1) but it used to be completely
ignored. We "fixed" this support in #2214, but to limit surprises on
upgrades from fed1, the schema upgrader had been modified to drop any
`@external` on an object type (since again, those were legal but ignored
in fed1). However, the code for adding @shareable in that same schema
upgrader was not updated correctly and so it _was_ taking `@external`
on types into account, and thus it ended up _not_ adding `@shareable`
in some cases where it should have.