Add missing input rewrites in some case of @interfaceObject (#2346)
Fetches to subgraphs that use `@interfaceObject` needs to have some
rewriting on their inputs to ensure we don't never send a __typename
whose value is an implementation type for the `@interfaceObject` (such
type is unknown to the particular subgraph).

The code was doing those rewrites in most cases, but in the case where
only fields from the `@interfaceObject` were requested _but_ some
type condition forced the resolution of the implementation, those
rewrites were not populated. It resulted in the subgraph rejecting
the `_entities` call due to some of the __typename not matching any
known type (by this subgraph).

This commit ensures those rewrites are correctly populated.

The commit also fix a minor typo in the result post-processing code.