Remove @apollo/server-gateway-interface
This interface never supported engines.node @ 12, so depending on
it here will cause the node 12 compatibility test to always fail.
Instead of depending on the package, I've copied the interface over
from the version we depended on. I don't expect we'll be implementing
any changes to the server/gateway interface for v0 versions of gateway
so this shouldn't cause any issues.