Adds debug query planner option to skip query planning for a single subgraph (#2441)
This commit adds a new option to the query planner config that, when
used and when the supergraph is built from only a single subgraph, skip
the query planning algorithm and instead generates a single fetch (to
the one subgraph) that directly uses the gateway "input" query.

This option is disabled by default and marked as a "debug" option as it
is a bit of a hack and so we're not suggesting committing to supporting
it forever. But it is easy enough to have now and can be convenient at
least for testing/debugging.

Note in particular that, when used, some of the features enabled by
the query planner (for instance `@defer`) cannot be enabled. The option
will also stop having any effect as soon as another subgraph is added,
not matter what that 2nd subgraph contains.