fix: Handle defaulted variables correctly during post-processing (#2443)
Post-processing didn't previously account for defaulted variable values in an
operation at all. This surfaces as a problem when using conditional directives
(@skip/@include) that use variables with defaulted values (and no
corresponding variable provided).

We should collect defaulted variable values from the operation and pass them
along with the variables we provide to computeResponse so these are
accounted for.