Restore backward compatible handling of corner case for fields (#2475)
It turns out that we have had (since fed 2.0 at least, maybe earlier) a
slightly weird behaviour (a bug really) in the validation of selection
sets which impacts what we accept for the `fields` arg of `@key`,
`@requires` and `@provides`. Namely, assuming a field `t` returnin an
object type, we were accepting a selection like:
```
{
  t {
    a
    b
  }
  t
}
```
even though the 2nd occurence of `t` is kind of incorrect according to
the graphQL spec since `t` should always have a sub-selection. But
the reason it works is that the code was merging the 1st and 2nd
occurrence of `t` before any validation was run, so internally the
selection is handled as just:
```
{
  t {
    a
    b
  }
}
```

Now, an assertion was added in #2387 that is triggered by the example
above, and that means that some `@key`, `@provides` or `@requires` that
were accepted (and were mostly correctly working) in currently released
versions would start erroring in 2.4 because of this.

To be clear, the historical behaviour is kind of wrong here, and we
should consider fixing it at some point. However, hard-failing on
upgrades is not very nice: we should probably introduce a warning
for a few versions before genuinely making this an error. Further,
the current assertion does not provide a very user friendly message.
In the meantime, this PR restore the status quo.