Fix assertion error due to not ignoring some unsatisfiable branches
Code was added in #2449 which "optimize" selections by removing
unecessary inline fragments and branches that can't be satisfied (when
we cross type conditions so that the intersection of types leading here
is empty).

Unfortunately that code misses some cases of when a branch is
impossible, which leads to trying to build a selection set that is
invalid and this in turn triggers an assertion error.