Fix incorrectly rejected composition and subgraph issues with `@interfaceObject` (#2494)
There is 2 main issues with the handling of `@interfaceObject` that this
patch fixes:
1. given some `@interfaceObject` type `I` in some subgraph `S1`, if
   another subgraph `S2` was declaring an _implementation_ type `T` of
   `I` but _without_ declaring `I`, then the code missed creating an
   "edge" between `T` in `S2` and `I` in `S1`, which resulted in
   composition rejecting such cases, even though it should work.
2. there were an additional issue in the handling of `@requires` so
   that the "input rewrites" generated were not generic enough, leading
   to sometimes sending incorrect queries to subgraph (more precisely,
   we were sometimes not rewriting the typename when querying an
   `@interfaceObject` subgraph, leading that subgraph to complain that
   it is asked for type it doesn't know).

Note that fixing those 2 issues surfaced the fact that the handling
of "rewrites" in the gateway was not working exactly as it should have.
More precisely, hen a path had a fragment `... on I`, if `I` was an
interface, then we would no select objects that implements `I`
correctly. As it happens, the router implementation of those rewrites
was both cleaner and didn't had that issue, so this patch also updates
the handling of "rewrites" to mimick what the router implementation
does (fixing the issue, and overall cleaning the code).

Fixes #2485