Fix assertion error on some multi-field `@requires` dependencies
When handling a `@requires`, the code gathers what it expected to be the set
of "fetch groups" needed to fulfill the requirement, but in the rare cases,
the same group can be added multiple times and in that case it was not
de-duplicated. This duplication was unexpected by the rest of the code,
ultimately leading to triggering an assertion.

This would happen when a `@requires` was on multiple fields, and one of those
field was also a condition (say a key) for another of the required field.