Fix assertion error with overlapping fragments and subtyping
When checking for name fragments to reuse, when some fields implementing
interfaces use subtyping (meaning that the type of the field
implementation is a sub-type of the type declared for that field in the
interface), an assertion error of the form `Cannot add selection of
field X to selection set of parent type Y` was sometimes raised (see
the comments in the commit for mode details).

Fixes #2592.