Fix possible fragment-related assertion error during query planning.
The assertion throw had message of the form `Cannot add fragment
of condition X (runtimes: ...) to parent type Y (runtimes: ...)` and
was due to not always properly maintaining the "parent type" information
when fragment spreads expanding into some other spread.

Additionally, this commit fixes a small issue in the code computing
the "diff" of what remains in a selection set after a fragment has
be "reused", which could lead to inefficient (and weird looking)
selections in fetches.