Memoize result of `FetchGroup.isUseless`
Profiling of some slow query planning shows `FetchGroup.isUseless`
as one of the hot path. This commit caches the result of this method
for a group, and only invalid that cache when we know the result may
needs to be recomputed. On the planning of some queries, this is
shown to provide a 15% improvement to query planning time.