Avoids considering keys that fetch themselves
When a type has multiple keys, the query planning was sometimes
considering an option where some key `x` was used to get field `y`
but then key `y` was used to get that same `y` field from another
subgraph. This is obviously not very useful, and we know we can
ignore those paths as the 1st part of those path already does what
we want. But considering those (useless) options, while harmless
for correction, was in some case drastically increasing the number
of plans that were evaluated, leading to long query planning
times.

In other words, this commit improve query planning times in some
cases (and at times significantly).