Renames/cleans up `trimUnsatisfiableBranches` method
The naming of `trimUnsatisfiableBranches` was misleading because it
does not just "remove unsatisfiable branches", it more generally remove
things that are not-very-optimal (even if they are not full branches).
So this rename this method to `normalize`, which more clearly explains
then intent (on top of being shorter).

Additionally, the method was not actually removing unsatisfiable
branches in all cases, which made it a bit error prone to use in
practice and led to some small inefficiencies in the
`NamedFragmentDefinition.selectionAtType` method. So fixed that,
making the method more regular and fixing the aforementioned
inefficiency.

This also force a normalisation after `expandAllFragments` because
we were relying on it in practice so this is less fragile this way.