Improve potential complexity of plan generation
When there is multiple valid choices for multiple paths of a query, the
number of theoretically possible ends up a cartesian product so it
explodes fairly quickly, and generating too many plans can be very
costly. To manage this in general, we generate the possible plans
incrementally and in an order that is likely to yield "good" plans
early, and use the cost of previously computed plan to possibly
cut early further computation.

The code doing this was however not optimal as we weren't using that
potential of early exit as aggressively as we could. This commit
fixes that.