Fix incorrect (too aggressive) optimisation
When building plans, after we've computed all the possible options to
get to a particular leaf of the query being plan, we're applying an
heuristic that checks if an option is guaranteed to be always better
than another one (and remove the latter one if so), which the goal
of reducing the number of actual full plan we need to evaluate.

But the implementation of that heuristic was incorrect: while it was
correctly removing options that we did meant to remove, it was also
sometimes removing options that really should be kept.

This resulted in planning sometimes not produce the most optimal
possible plan.

This commit fixes that heuristic to only exclude what can be safely
excluded.