Avoid leaving unused fragments
Some interaction of the code that re-expand named fragments that are
used only once and the normalization we code post expansion could leave
some fragment unused, yet included in the list of fragment, which
ultimately led to an invalid query. This commit make sure we don't
include such unused fragments.