Fix incorrect removal of fragments only used by other fragments (#2648)
The patch from #2628, while it did fix the issue of leaving unused
fragments, mistakenly introduced a regression in that fragments
that are used by other fragments but are not used in the query
selection are now removed, which is obviously undesirable.

This was simply because the patch of #2628 only counted the usages
in the main selection, not the ones other fragments as it should
have. This commit fixes that.