Skip supergraph validations in production by default
When receiving a supergraph, we were always running all the graphQL
validations on both the supergraph itslef, and on each of the subgraphs
we extract from it.

This is useful for us when doing dev, as it allows to get decent error
message when we break the supergraph generation process or the subgraph
extraction process, but this is of very limited value otherwise: even if
we don't run those full validations, the supergraph still has to be
valid graphQL syntax to parse in the first place, and the subgraph
extraction will still throw if the supergraph is not essentially a valid
supergraph (and note that the code checking that the `join` spec version
is valid, or that the supergraph does not use non-supported feature for
`SECURITY` or `EXECUTION` is *not* disabled by this).

Besides, it's not like uplink will send basically-valid-but-slightly-corrupted
supergraphs, and when using `IntrospectAndCompose` there isn't really
room for the supergraph to be corrupted.

And those validations are actually not super cheap, especially on large
supergraphs, so running them feels wasteful.

This commit thus disable those validation in production environment
(when `NODE_ENV === "production"`) by default. It keeps those validation
during dev mainly for 2 reasons:
1. as mentioned above, so that those validate are run when we run our
  own unit tests, to help catch/identify any subtle mistake we would
  introduce in the supergraph generation and extaction codes.
2. when debugging something, it can happens that one start a gateway
  from a statically provided supergraph SDL, SDL that one may have
  manually edited for some reason (testing something, playing with it,
  ...). In that case, the validation can provide nicer errors if
  the manual edits are invalid.

The commit also introduce a `validateReceivedSupergraphs` in the
gateway config to allows overriding the default behaviour described
above.