Fix normalization of fragment spreads
The code that "normalize" selection sets to remove unecessary fragments
is also used to remove impossibe (and thus invalid) branches when some
reused fragments gets expanded away due to only be used once.

However, while the code for inline fragments was correctly removing
impossible branches, the similar code for named spreads was not, which
in some rare cases could result in an invalid subgraph fetch.