Introduce `@requiresScopes` directive in composition (#2649)
With this change, users can now compose `@requiresScopes`
applications from their subgraphs into a supergraph. This addition
will support a future version of Apollo Router that enables
authenticated access to specific types and fields via directive
applications.

Since the implementation of `@requiresScopes` is strictly a
composition and execution concern, there's no change to the
query planner with this work. The execution work is well under way
in Apollo Router (and won't be built at all for Gateway). So as far as
this repo is concerned, only composition is concerned with
`@requiresScopes`.

Co-authored-by: Sylvain Lebresne <lebresne@gmail.com>