Fix potential normalization assertion error for fragment on abstract types (#2725)
An abstract type (interface or union) can have different runtime types
in different subgraphs. For instance, a type `T` may implement some
interface `I` in subgraph A, but not in subgraph B _even_ if `I` is
otherwise defined in subgraph B (admittedly something not to be abused,
but it can be convenient as a temporary state when evolving schema).

For named fragment, this can create a case where a fragment on an
abstract type can be "rebased" on a subgraph, but where some of its
application (spread) in other fragments are invalid due to being
use in the context of a type that intersect the abstrat type in the
supergraph but not in the particular subgraph. When that's the case,
the invalid spread (for the subgraph) needs to be expanded, and that
expansion properly "rebased", but the code was not handling that case
at all. Instead it kept the invalid spread, and this led to some
later assertion errors.

Fixes #2721.