Fix incorrect removal of dependencies between groups in some rare `@requires` (#2726)
Fulfilling some `@requires` can necessitate querying group(s)
(subgraphs) that are not the direct parent of the group with the
`@requires`. To generate plan that are as optimal as possible, the code
needs to figure out the proper parent of those new groups, that is
what is the "earliest" group those new groups truly depend on. This
is done by starting to look at the parent of the `@requires` group,
see if the new groups genuinely depend on anything that parent fetches,
and if not, iterate the same logic on the grand-parent (until we either
find a parent the new groups depend on, or we reach the beginning of
the plan.

But the code that checked if the new groups truly depend on a given
parent was incomplete: it was sometimes responding that there were
no dependency when that was not true.

The result is that for some specific configuration of a `@requires`,
we could end up with a plan whereby some group/fetch A depended on
the results of another group/fetch B, but A was performed in parallel
of B instead of after it. Which ultimately resulted in not getting
the proper data for the required fields, leading to incorrect executions
(the exact detail of how this would show up depends a bit on the exact
service implementation, but typically the resolver of the field on
which there is a `@requires` would get `null` for its requirements,
which might lead that resolver to throw an error as this is unexpected).

Fixes #2683

Co-authored-by: Chris Lenfest <clenfest@apollographql.com>