Fix `__typename` rebasing for interface objects (#2886)
We should never query `__typename` from a subgraph for an object type
marked `@interfaceObject`, as the value we get back will always be
wrong. Normally, we prevent this by not having a `__typename` edge in
the federated query graph.

However, during the optimization where we try to reuse existing named
fragments, we rebase them from the API schema onto the subgraph schemas.
We usually ignore selections where this leads to invalidity (e.g. fields
or types that don't exist in the subgraph schema), but the code doesn't
currently view the `__typename` field being rebased onto an interface
object as invalid. This has led to bugs where named fragments containing
`__typename` accidentally cause it to be queried on interface objects in
subgraph queries.

This PR changes `Field.rebaseOn()` such that `__typename` being rebased
onto a parent type that is an interface object is considered invalid.
Additionally, if the parent type is an abstract type, and that abstract
type could possibly be an interface object type at runtime, then this is
additionally considered invalid.